Implement a --get-server-output flag that allows the client to retrieve (most of) the output emitted by the server.
If the server was invoked with the --json flag, the output will be in
JSON, otherwise it will be in the human-readable format.

If the client was invoked with the --json flag, the output will be
contained within the JSON output structure, otherwise it will be
appended (in whatever format) to the bottom of the human-readable
output.

Because of the sequencing of the output generation and display, the
server-side output includes only the starting output, interval
statistics and summaries, but not the overall summaries.  (The overall
summaries were already displayed in the client's output.)

Towards issue #160.

(cherry picked from commit ba8d6e6246d5e6d7a4417e6b0990846f89a00a90)
Signed-off-by: Bruce A. Mah <bmah@es.net>

Conflicts:
	src/iperf3.1
	src/iperf_api.c
	src/iperf_api.h