Only do the IPV6_V6ONLY setsockopt(3) stuff if that option is available.
Should fix #177, in which compilation failed on older Solaris systems
that didn't have it.  This is a different approach than a patch
suggested in that issue.

Weakly regression-tested on other platforms (test this by specifying
-6, -4, or neither to the server when binding to the wildcard address,
and seeing if a client can connect with various of -6, -4, or neither).

(cherry picked from commit 147d3369a0a9f448875088915e74a284a46436ab)
Signed-off-by: Bruce A. Mah <bmah@es.net>

Conflicts:
	src/iperf_sctp.c