Fix out-of-the-box builds with no OpenSSL (#633)
Don't make OpenSSL detection error out unless --with-openssl explicitly given.

This fixes the out-of-the-box build on macOS, which doesn't include
headers, etc. for OpenSSL by default.  The configure script still
errors out if the user specifies a --with-openssl flag but for some
reason the build host's OpenSSL is broken or can't be used.

* Regen.