Close buffer_fd after use with mmap(), to allow more concurrent streams
After calling mmap() the file descriptor of the underlying file is no
longer needed, and closing that file descriptor doubles the maximum
number of concurrent streams.

On a Linux system with 1024 file descriptors per process, this change
will allow up to 1016 concurrent streams.

This change also makes the maximum stream count as a function of
per-process file descriptor limit heuristic implemented in the previous
patch work correctly. In other words, either the requested -P count is
acomodated, or a usage error is issued.