set TCP_NODELAY flag for ctrl socket on server side
in scenarios with high packet loss rate, for example due
to overflowing the capacity of the channel, the TCP traffic
on the control socket might be delayed too much. In a
particular case with a high-bandwidth bi-directionl UDP
test that exceeded the channel capacity by far, the UDP traffic
generated at the server was already causing full tx queues.
this caused the "delayed" TCP control packet to the client to be
also dropped. thus, the client never entered the TEST_RUNNING state.

this patch sets the TCP_NODELAY flag on the server's control
socket and makes sure that the control traffic goes through the
client before the UDP stream is even started.