iperf_udp: implement UDP connect retry mechanism
Sometimes the single UDP datagram sent as a form of "connection init"
gets lost on the way. A common example is testing over wifi, or more
specifically in our case testing many parallel udp connections.

When this happens the recv() call on the client times out after 30
seconds and the test is unsuccessful.

This commit implements a very basic retry mechanism for this. Using
select() calls with a timeout of 1 second we attempt to send the
"connection init" datagram up to 5 times and expect a reply from the
server.

If this fails the test fails earlier reporting issues with initializing
the connection. Otherwise the test proceeds as before.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>